backend/requirements.txt
flask==2.0.1
werkzeug==2.0.3
flask-cors==3.0.10
pandas==1.3.5
numpy==1.21.6
scikit-learn==1.0.2
xgboost==1.5.2
joblib==1.1.0
requests==2.27.1
python-dotenv==0.20.0